Added the Scanning Station Selector
Depending upon which documents are marked off on the check-list when
students arrive, they can be sent to either the scanning station (those
who need to begin from scratch) or to the I-94 retrieval station (those
who have already submitted all their documents to iHawk except for
their I-94). This new addition is a page that lets the instructor get
the correct set of modules for the student at their station.