Added the I-94 Short Circuit
For the students at the “I-94 Station”, it is understood that they have
previously already submitted their valid passport, admission document
respective to their category (such as I-20 for F-1/F-2 and DS-2019 for
J-1/J-2) and their most recent visa used at the port of entry in the
US. When logged into iHawk all of the upload forms for the
aforementioned documents are skipped and the I-94 upload portal is
opened.

Students at the I-94 Retrieval and Upload Station still have to visit
and update their Emergency Contact Information forms and Local Address
Information forms.

[Students J-1 and J-2 categories also need to submit their local
contact number before going to the Administrative Approval Page.]