Another Approach : Select Completed
Earlier, we asked the students to determine the forms that applied to
them and then to select whichever ones they still needed to submit.

Now, instead, the students will indicate which forms they have already
submitted. Based on their selection, their category can be determined
and it can be figured out which forms they still need to submit.